CXP-965: refactor DeleteAppHandlerIntegration test
