+++ src/Akeneo/Connectivity/Connection/back/tests/CatalogBuilder/Resources/config/loaders.yml
@@ -105,3 +105,12 @@ services:
         class: 'Akeneo\Connectivity\Connection\Tests\CatalogBuilder\ConnectedAppLoader'
         arguments:
             - '@database_connection'
+            - '@akeneo_connectivity.connection.fixtures.enrichment.user_role_loader'
+            - '@fos_oauth_server.storage'
+            - '@akeneo_connectivity.connection.service.apps.client_provider'
+            - '@akeneo_connectivity.connection.service.user.create_user_group'
+            - '@akeneo_connectivity.connection.service.user.create_user'
+            - '@Akeneo\Connectivity\Connection\Application\Apps\Service\CreateConnection'
+            - '@Akeneo\Connectivity\Connection\Application\Apps\Service\CreateConnectedApp'
+            - '@pim_user.repository.user'
+            - '@pim_connector.doctrine.cache_clearer'
